requirements.txt
futures==3.0.5
Jinja2==2.8
PyYAML==3.11
requests==2.9.1
sh==1.11
tornado==4.3
docopt==0.6.2
schema==0.5.0

#
# Dev
#
mock
nose
pep8
pyflakes
coverage
